Fixed GPS only bug. Added BaseLocationClient abstract class.
